Cast rr stuff to ints because its dumb

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/roadrunner/TwoDeadWheelLocalizer.java
@@ -90,8 +90,8 @@ public final class TwoDeadWheelLocalizer implements Localizer {
         if (!initialized) {
             initialized = true;
 
-            lastParPos = parPosVel.position;
-            lastPerpPos = perpPosVel.position;
+            lastParPos = (int) parPosVel.position;
+            lastPerpPos = (int) perpPosVel.position;
             lastHeading = heading;
 
             return new Twist2dDual<>(
@@ -100,8 +100,8 @@ public final class TwoDeadWheelLocalizer implements Localizer {
             );
         }
 
-        int parPosDelta = parPosVel.position - lastParPos;
-        int perpPosDelta = perpPosVel.position - lastPerpPos;
+        int parPosDelta = (int) parPosVel.position - lastParPos;
+        int perpPosDelta = (int) perpPosVel.position - lastPerpPos;
         double headingDelta = heading.minus(lastHeading);
 
         Twist2dDual<Time> twist = new Twist2dDual<>(
@@ -121,8 +121,8 @@ public final class TwoDeadWheelLocalizer implements Localizer {
                 })
         );
 
-        lastParPos = parPosVel.position;
-        lastPerpPos = perpPosVel.position;
+        lastParPos = (int) parPosVel.position;
+        lastPerpPos = (int) perpPosVel.position;
         lastHeading = heading;
 
         return twist;
